Reorganize input and selector tests
