--- tests/test_input_dispatch.py
@@ -1,5 +1,0 @@
-# ruff: noqa: F401,F403
-
-from .input_dispatch_browsing_cases import *  # noqa: F401,F403
-from .input_dispatch_dialogs_cases import *  # noqa: F401,F403
-from .input_dispatch_palette_cases import *  # noqa: F401,F403
